Added examples + test
Added an example of how to use the library. Its not the best dataset to
use, so I'll need to find a better one. Implementing this example lead
to the creation of a new test and the breaking of all others. I need to
find a way to check for classes being added via d3 (or scrap the
auto-adding of classes all together).